requirements.txt
Flask==3.1.1
mysql-connector-python==9.3.0
requests==2.32.4

